requirements.txt
Flask==3.0.3
boto3==1.34.162
botocore==1.34.162
authlib==1.3.0
python-dotenv==1.0.1
gunicorn==22.0.0
Werkzeug==3.0.3